Extracted parseValidatorEvent to commons
