+++ monitor/utils/validatorUtils.js
@@ -1,33 +1,8 @@
 /* eslint no-param-reassign: ["error", { "props": false }] */
 
-const { BRIDGE_VALIDATORS_ABI } = require('../../commons')
+const { BRIDGE_VALIDATORS_ABI, parseValidatorEvent } = require('../../commons')
 const logger = require('../logger')('validatorsUtils')
 const { getPastEvents } = require('./contract')
-
-const parseValidatorEvent = event => {
-  if (
-    event.event === undefined &&
-    event.raw &&
-    event.raw.topics &&
-    (event.raw.topics[0] === '0xe366c1c0452ed8eec96861e9e54141ebff23c9ec89fe27b996b45f5ec3884987' ||
-      event.raw.topics[0] === '0x8064a302796c89446a96d63470b5b036212da26bd2debe5bec73e0170a9a5e83')
-  ) {
-    const rawAddress = event.raw.topics.length > 1 ? event.raw.topics[1] : event.raw.data
-    const address = '0x' + rawAddress.slice(26)
-    event.event = 'ValidatorAdded'
-    event.returnValues.validator = address
-  } else if (
-    event.event === undefined &&
-    event.raw &&
-    event.raw.topics &&
-    event.raw.topics[0] === '0xe1434e25d6611e0db941968fdc97811c982ac1602e951637d206f5fdda9dd8f1'
-  ) {
-    const rawAddress = event.raw.data === '0x' ? event.raw.topics[1] : event.raw.data
-    const address = '0x' + rawAddress.slice(26)
-    event.event = 'ValidatorRemoved'
-    event.returnValues.validator = address
-  }
-}
 
 const processValidatorsEvents = events => {
   const validatorList = new Set()
